0807修改版本2.18.0的bug

+++ OpenEdXMobile/src/main/java/org/edx/mobile/model/api/CourseEntry.java
@@ -225,10 +225,10 @@ public class CourseEntry implements Serializable {
     }
 
     public static class Progress implements Serializable{
-        float total_grade;
+        double total_grade;
         boolean is_pass;
 
-        public float getTotal_grade() {
+        public double getTotal_grade() {
             return total_grade;
         }
